remove awserr
